bookHubBack/requirements.txt
djoser==2.2.2
djangorestframework_simplejwt==5.3.1
djangorestframework==3.14.0
django==4.2
django-cors-headers==4.3.1
dj-rest-auth==4.0.0
pillow==9.5.0
scikit-learn
django-filter
pandas
django-csp
gunicorn
